worker/requirements.txt
fastapi
uvicorn
elasticsearch
elastic-opentelemetry
boto3>1.34
redis>=5.0,<6.0
python-multipart          # FastAPI file uploads
Pillow
exifread                  # or 'piexif', either works
numpy
opencv-python-headless
torch                     # cpu ok; big, but works
open-clip-torch           # CLIP embeddings
orjson
rq                        # job queue
open_clip_torch
aioredis
aiohttp
requests
redis>=5.0